refactor: Update factories to use existing users and roles; adjust user role in UserFactory and seeder for consistency

+++ database/migrations/2025_07_07_143551_remove_columns_from_orders_table.php
@@ -23,8 +23,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('orders', function (Blueprint $table) {
+            if (Schema::hasColumn('orders', 'product_id')) {
             $table->foreignId('product_id')->constrained('products')->onDelete('restict');
             $table->unsignedInteger('quantity');
+            }
         });
     }
 };
